Refactor the examples
Move them from DSx directory up
