+++ examples/tracing/retrieve-query-trace.js.broken
