+++ examples/tuple/tuple-insert-select.js
