+++ examples/udt/udt-insert-select.js
