+++ examples/util.js
